Simplify: extract shared findAndInvoke helper, add ancestor walk depth limit
- Extract duplicated find-by-label + ancestor-walk logic from tap_element
  and long_press into FIND_AND_INVOKE_JS in fiber.ts
- Add depth limit (50) to ancestor walk to prevent runaway traversal
- Remove narration comments; remove misleading 'pressable' variable name
